updated licence and org migration script
